# Simplified benchmark output with --benchmark-no-analysis