Removed unused informations from benchmarks output when --benchmark-no-analysis flag is used to improve readability.