@RT222 2019-10-07
Before :
```
...............................................................................

benchmark name                                  samples       iterations    estimated
                                                mean          low mean      high mean
                                                std dev       low std dev   high std dev
-------------------------------------------------------------------------------
1 component creation                                      100             1    64.1466 ms
                                                   234.586 us    234.586 us    234.586 us
                                                         0 us          0 us          0 us

3 components creation                                     100             1    83.6913 ms
                                                   315.606 us    315.606 us    315.606 us
                                                         0 us          0 us          0 us

10 components creation                                    100             1    128.337 ms
                                                   663.002 us    663.002 us    663.002 us
                                                         0 us          0 us          0 us

```
After :
```
...............................................................................

benchmark name                            samples    iterations          mean
-------------------------------------------------------------------------------
1 component creation                           100             1    234.586 us
3 components creation                          100             1    315.606 us
10 components creation                         100             1    663.002 us
```